--- a/src/mjolnir/pbfadminparser.cc
+++ b/src/mjolnir/pbfadminparser.cc
@@ -104,7 +104,7 @@ public:
     admin.roles.reserve(members.size());
     for (const auto& member : members) {
 
-      if (member.member_type == OSMPBF::Relation::MemberType::Relation_MemberType_WAY) {
+      if (member.member_type == OSMPBF::Relation::MemberType::WAY) {
         members_.set(member.member_id);
         admin.ways.push_back(member.member_id);
         admin.roles.push_back(member.role != "inner"); // assume outer
@@ -148,34 +148,24 @@ OSMAdminData PBFAdminParser::Parse(const boost::property_tree::ptree& pt,
 
   LOG_INFO("Parsing files: " + boost::algorithm::join(input_files, ", "));
 
-  // hold open all the files so that if something else (like diff application)
-  // needs to mess with them we wont have troubles with inodes changing underneath us
-  std::list<std::ifstream> file_handles;
-  for (const auto& input_file : input_files) {
-    file_handles.emplace_back(input_file, std::ios::binary);
-    if (!file_handles.back().is_open()) {
-      throw std::runtime_error("Unable to open: " + input_file);
-    }
-  }
-
   // Parse each input file for relations
   LOG_INFO("Parsing relations...");
-  for (auto& file_handle : file_handles) {
-    OSMPBF::Parser::parse(file_handle,
-                          static_cast<OSMPBF::Interest>(OSMPBF::Interest::RELATIONS |
-                                                        OSMPBF::Interest::CHANGESETS),
-                          callback);
+  for (auto& file : input_files) {
+    OSMPBF::parse(file,
+                  static_cast<OSMPBF::Interest>(OSMPBF::Interest::RELATIONS |
+                                                OSMPBF::Interest::CHANGESETS),
+                  callback);
   }
   LOG_INFO("Finished with " + std::to_string(osmdata.admins.size()) +
            " admin polygons comprised of " + std::to_string(osmdata.osm_way_count) + " ways");
 
   // Parse the ways.
   LOG_INFO("Parsing ways...");
-  for (auto& file_handle : file_handles) {
-    OSMPBF::Parser::parse(file_handle,
-                          static_cast<OSMPBF::Interest>(OSMPBF::Interest::WAYS |
-                                                        OSMPBF::Interest::CHANGESETS),
-                          callback);
+  for (auto& file : input_files) {
+    OSMPBF::parse(file,
+                  static_cast<OSMPBF::Interest>(OSMPBF::Interest::WAYS |
+                                                OSMPBF::Interest::CHANGESETS),
+                  callback);
   }
   LOG_INFO("Finished with " + std::to_string(osmdata.way_map.size()) + " ways comprised of " +
            std::to_string(osmdata.node_count) + " nodes");
@@ -183,11 +173,11 @@ OSMAdminData PBFAdminParser::Parse(const boost::property_tree::ptree& pt,
   // Parse node in all the input files. Skip any that are not marked from
   // being used in a way.
   LOG_INFO("Parsing nodes...");
-  for (auto& file_handle : file_handles) {
-    OSMPBF::Parser::parse(file_handle,
-                          static_cast<OSMPBF::Interest>(OSMPBF::Interest::NODES |
-                                                        OSMPBF::Interest::CHANGESETS),
-                          callback);
+  for (auto& file : input_files) {
+    OSMPBF::parse(file,
+                  static_cast<OSMPBF::Interest>(OSMPBF::Interest::NODES |
+                                                OSMPBF::Interest::CHANGESETS),
+                  callback);
   }
   LOG_INFO("Finished with " + std::to_string(osmdata.osm_node_count) + " nodes");