… CHANGELOG.md
… proto/CMakeLists.txt
… src/mjolnir/CMakeLists.txt
… test/astar.cc
… test/gurka/gurka.cc
… test/gurka/test_admin.cc
… test/gurka/test_admin_sidewalk_crossing_override.cc
… test/gurka/test_admin_uk_override.cc
… test/gurka/test_build_admin.cc
… test/gurka/test_elevation.cc
… test/gurka/test_landmarks.cc
… test/gurka/test_languages.cc
… test/gurka/test_phonemes.cc
… test/gurka/test_phonemes_w_langs.cc
… test/gurka/test_stop_signs.cc
… test/gurka/test_traffic_signals.cc
… test/gurka/test_yield_signs.cc
… valhalla/mjolnir/util.h