# Issue

resolves https://github.com/valhalla/valhalla/issues/219
also closes https://github.com/valhalla/valhalla/issues/1462 by removing raw pointers from osmpbfparser
and closes https://github.com/valhalla/valhalla/issues/3900 by removing own coordinates decoding for dense nodes

Speed up pbf parsing by using libosmium, that Valhalla already has as a dependency.

Associated functional changes:
- **Parsing directly from URL** - with libosmium it is possible to build admins and tileset directly from remote file like `./valhalla_build_admins -c ./config.json https://download.geofabrik.de/europe/sweden-latest.osm.pbf`. Ofc. current flow is not optimised for such flow and file will be downloaded 3 times (as Valhalla parses pbf in 3 passes).
- **No more pbf pinning** - sadly, it is not possible to "pin" file with libosmium (libosmium opens file and starts reader thread [in the same place](https://github.com/osmcode/libosmium/blob/master/include/osmium/io/reader.hpp#L319-L350)), so if during the parse phase the pbf file was moved or removed then the build process will fail.

A note about a2c24c48d9144f25205c57c324d2f74a5fe59014 - during the final checks with [profiler](https://github.com/mstange/samply/) I've noticed that quite a big amount of time is spent on construction/destruction of the `std::unordered_map` that is used to store node/way/relation tags (see `Tags`). Replacing it by `robin_hood` hash map noticeably improved `valhalla_build_admins` times, 70s -> 55s for Germany on my m3 max laptop.

## Benchmarks

On [m7a.2xlarge EC2 instance](https://aws.amazon.com/ec2/instance-types/) (8 CPUs, 32 GB RAM) with 1000 MB/s throughput 12k IOPS disk

| Operation | master | this PR |
| --- | --- | --- |
| `valhalla_build_admins` | 2h 20m | 39m |
| `valhalla_build_tiles --end parsenodes` | 8h 33m | 4h 52m |

Script to reproduce

```sh
git clone --recurse-submodules https://github.com/valhalla/valhalla.git
#git clone --recurse-submodules https://github.com/kinkard/valhalla.git --branch parse-using-libosmium

cd valhalla

./scripts/install-linux-deps.sh 

mkdir build && cd build
cmake .. -DCMAKE_BUILD_TYPE=Release -DENABLE_SINGLE_FILES_WERROR=OFF
cmake --build . -j 8

sudo snap install aws-cli --classic
aws s3 --no-sign-request --no-progress cp s3://osm-planet-eu-central-1/planet/pbf/2025/planet-250120.osm.pbf ./

mkdir valhalla_data

./valhalla_build_config \
--mjolnir-tile-dir=./valhalla_data \
--mjolnir-tile-extract=./valhalla_data/tiles.tar \
--mjolnir-traffic-extract=./valhalla_data/traffic.tar \
--mjolnir-admin=./valhalla_data/admins.sqlite \
--mjolnir-timezone=./valhalla_data/timezones.sqlite \
> ./valhalla_data/config.json

./valhalla_build_timezones > ./valhalla_data/timezones.sqlite

time ./valhalla_build_admins -c ./valhalla_data/config.json *.osm.pbf

time ./valhalla_build_tiles -c ./valhalla_data/config.json --end parsenodes *.osm.pbf
```

<details>
<summary>Full build tiles log from master</summary>

```
2025/01/25 16:04:05.141412 [INFO] Running valhalla_build_tiles with 8 thread(s).
2025/01/25 16:04:05.141460 [INFO] Start stage = initialize End stage = parsenodes
2025/01/25 16:04:05.150769 [INFO] Parsing files for ways: planet-250120.osm.pbf
2025/01/25 16:04:05.151927 [INFO] Parsing ways...
2025/01/25 16:25:36.011165 [INFO] invalid_argument thrown for condition (2024 Jul 01- 2024 Sep 13)
2025/01/25 16:25:36.015004 [INFO] invalid_argument thrown for condition (2024 Jul 01- 2024 Sep 13)
2025/01/25 16:26:25.473709 [INFO] invalid_argument thrown for condition (Jul 1- Oct 1)
2025/01/25 16:29:25.138732 [INFO] invalid_argument thrown for condition (2023 Dec 31- 2024 Oct 25)
2025/01/25 16:29:28.377766 [INFO] invalid_argument thrown for condition (2024 May 26- 2024 Aug 30)
2025/01/25 16:31:29.803795 [INFO] invalid_argument thrown for condition (2024 Mar 10- 2025 Feb 28)
2025/01/25 16:31:58.840200 [WARN] Error during parsing of `layer` tag on the way 31730471: stoi
2025/01/25 16:31:58.840285 [WARN] Error during parsing of `layer` tag on the way 31730472: stoi
2025/01/25 16:33:08.819604 [WARN] Error during parsing of `layer` tag on the way 35744549: stoi
2025/01/25 16:33:33.401001 [INFO] invalid_argument thrown for condition (2023 Nov 21- 2024 Aug 05)
2025/01/25 16:33:33.639493 [INFO] invalid_argument thrown for condition (2024 Apr 13- 2024 Jul 12)
2025/01/25 16:33:38.176519 [INFO] invalid_argument thrown for condition (2024 Mar 10- 2025 Feb 28)
2025/01/25 16:33:40.120248 [WARN] Error during parsing of `layer` tag on the way 37975571: stoi
2025/01/25 16:35:44.015324 [WARN] Error during parsing of `layer` tag on the way 47287106: stoi
2025/01/25 16:37:58.919001 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:37:58.919911 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:37:58.920137 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:37:58.921700 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.948050 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.969675 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.970067 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.971608 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.971812 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.972902 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.974697 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.975129 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.975295 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.976039 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.979567 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:38:38.979703 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:42:24.040237 [WARN] Exceeded max speed for way id: 101621149
2025/01/25 16:42:24.040277 [WARN] Exceeded max speed for way id: 101621149
2025/01/25 16:42:42.771166 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:44:43.649209 [INFO] invalid_argument thrown for condition (2024 Jul 01- 2024 Aug 30)
2025/01/25 16:45:13.613636 [INFO] invalid_argument thrown for condition (Dec 27- Feb 29: Fr 19:00-24:00 "Calle recreativa")
2025/01/25 16:46:11.887667 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:11.887829 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:11.919481 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:11.919652 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:11.973542 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:11.975410 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:11.976553 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.179809 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.187335 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.187467 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.187622 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.224831 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.225174 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.226444 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.232331 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.233455 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.233636 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.268845 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.268980 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:46:12.271236 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 16:47:54.326604 [INFO] invalid_argument thrown for condition Jul-Sep 09:30-16:00 "online reservation only"
2025/01/25 16:49:44.258400 [INFO] invalid_argument thrown for condition (2024 Jul 01- 2024 Aug 30)
2025/01/25 16:52:58.691327 [WARN] Exceeded max speed for way id: 166291289
2025/01/25 16:58:47.810836 [WARN] Exceeded max speed for way id: 199564443
2025/01/25 16:58:47.810874 [WARN] Exceeded max speed for way id: 199564443
2025/01/25 17:07:37.386817 [INFO] invalid_argument thrown for condition (2024 Mar 10- 2024 May 31)
2025/01/25 17:08:31.734602 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 17:08:54.339174 [INFO] invalid_argument thrown for condition (Dec 27- Feb 29: Fr 19:00-24:00 "Calle recreativa")
2025/01/25 17:16:15.757566 [WARN] Error during parsing of `layer` tag on the way 314564421: stoi
2025/01/25 17:22:51.525972 [INFO] invalid_argument thrown for condition (Jun 15-Labor Day)
2025/01/25 17:28:41.575494 [INFO] invalid_argument thrown for condition (Jun 15-Labor Day)
2025/01/25 17:32:11.198765 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 17:37:44.096217 [INFO] invalid_argument thrown for condition (2023 Dec 31- 2024 Oct 25)
2025/01/25 17:52:00.483218 [INFO] invalid_argument thrown for condition (Nov-Mar 08:00-17:00, Apr-Oct 08:00-18:30)
2025/01/25 17:57:47.968380 [WARN] Exceeded max speed for way id: 625497422
2025/01/25 17:57:47.968459 [WARN] Exceeded max speed for way id: 625497423
2025/01/25 17:57:48.195376 [WARN] Exceeded max speed for way id: 625523473
2025/01/25 17:57:58.837266 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:00:03.231638 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:00:03.462847 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:00:03.479345 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:00:05.099906 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:00:05.112673 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:06:26.604303 [WARN] Exceeded max speed for way id: 687203695
2025/01/25 18:06:26.604342 [WARN] Exceeded max speed for way id: 687203695
2025/01/25 18:09:32.733048 [WARN] Exceeded max speed for way id: 706566227
2025/01/25 18:09:32.733119 [WARN] Exceeded max speed for way id: 706566228
2025/01/25 18:13:16.889836 [WARN] Error during parsing of `layer` tag on the way 735049192: stoi
2025/01/25 18:13:43.780349 [WARN] Exceeded max speed for way id: 738321405
2025/01/25 18:13:43.780424 [WARN] Exceeded max speed for way id: 738321406
2025/01/25 18:13:43.780587 [WARN] Exceeded max speed for way id: 738321412
2025/01/25 18:13:43.780628 [WARN] Exceeded max speed for way id: 738321413
2025/01/25 18:13:43.780737 [WARN] Exceeded max speed for way id: 738321416
2025/01/25 18:13:43.780776 [WARN] Exceeded max speed for way id: 738321417
2025/01/25 18:13:43.780811 [WARN] Exceeded max speed for way id: 738321418
2025/01/25 18:13:43.780834 [WARN] Exceeded max speed for way id: 738321419
2025/01/25 18:18:00.245783 [WARN] Error during parsing of `layer` tag on the way 769745481: stoi
2025/01/25 18:23:51.014979 [INFO] invalid_argument thrown for condition (Dec 26-easter -2 days)
2025/01/25 18:26:58.144014 [INFO] invalid_argument thrown for condition (Nov-Mar 08:00-17:00, Apr-Oct 08:00-18:30)
2025/01/25 18:27:41.091515 [INFO] invalid_argument thrown for condition (2024 Mar 04- 2024 Dec 24)
2025/01/25 18:36:35.493670 [WARN] Exceeded max speed for way id: 926857769
2025/01/25 18:41:24.994626 [WARN] Exceeded max speed for way id: 968035542
2025/01/25 18:41:24.994654 [WARN] Exceeded max speed for way id: 968035542
2025/01/25 18:42:10.083854 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:42:10.084160 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:42:10.084290 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:33.831703 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.430545 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.430695 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.430803 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.430917 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.431032 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.431141 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.431249 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.454187 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:49:37.454341 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 18:50:19.358047 [WARN] Exceeded max speed for way id: 1050272991
2025/01/25 18:50:19.358233 [WARN] Exceeded max speed for way id: 1050272992
2025/01/25 18:50:19.358274 [WARN] Exceeded max speed for way id: 1050272993
2025/01/25 18:50:19.358314 [WARN] Exceeded max speed for way id: 1050272994
2025/01/25 18:50:19.358343 [WARN] Exceeded max speed for way id: 1050272995
2025/01/25 18:52:50.462837 [WARN] Exceeded max speed for way id: 1073024251
2025/01/25 18:52:50.462873 [WARN] Exceeded max speed for way id: 1073024251
2025/01/25 18:54:37.003163 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:54:37.004382 [INFO] invalid_argument thrown for condition  Apr-Sep Su[1] 11:00-18:00)
2025/01/25 18:57:51.890583 [INFO] invalid_argument thrown for condition (Apr-Sep 20:00-06:00, Oct-Mar 18:00-06:00)
2025/01/25 19:02:48.202124 [INFO] invalid_argument thrown for condition (2023 May 01- 2024 Aug 30)
2025/01/25 19:09:23.498566 [WARN] Error during parsing of `layer` tag on the way 1229020034: stoi
2025/01/25 19:11:39.851352 [WARN] Error during parsing of `layer` tag on the way 1250075792: stoi
2025/01/25 19:12:57.972540 [WARN] Error during parsing of `layer` tag on the way 1264574506: stoi
2025/01/25 19:13:09.636101 [WARN] Error during parsing of `layer` tag on the way 1266585985: stoi
2025/01/25 19:15:57.694893 [WARN] Error during parsing of `layer` tag on the way 1295711042: stoi
2025/01/25 19:16:08.895410 [WARN] Error during parsing of `layer` tag on the way 1297388709: stoi
2025/01/25 19:16:13.565358 [WARN] Error during parsing of `layer` tag on the way 1298158378: stoi
2025/01/25 19:18:13.443844 [WARN] Error during parsing of `layer` tag on the way 1314553538: stoi
2025/01/25 19:18:36.538696 [WARN] Error during parsing of `layer` tag on the way 1317496363: stoi
2025/01/25 19:19:16.432929 [WARN] Error during parsing of `layer` tag on the way 1323775572: stoi
2025/01/25 19:19:16.433093 [WARN] Error during parsing of `layer` tag on the way 1323775577: stoi
2025/01/25 19:19:16.433148 [WARN] Error during parsing of `layer` tag on the way 1323775579: stoi
2025/01/25 19:19:16.433196 [WARN] Error during parsing of `layer` tag on the way 1323775581: stoi
2025/01/25 19:19:16.433243 [WARN] Error during parsing of `layer` tag on the way 1323775583: stoi
2025/01/25 19:19:16.433289 [WARN] Error during parsing of `layer` tag on the way 1323775585: stoi
2025/01/25 19:20:20.768424 [WARN] Error during parsing of `layer` tag on the way 1333729160: stoi
2025/01/25 19:20:43.941295 [WARN] Error during parsing of `layer` tag on the way 1337494001: stoi
2025/01/25 19:22:02.663802 [WARN] Error during parsing of `layer` tag on the way 1348919494: stoi
2025/01/25 19:43:04.277249 [INFO] Added 344752 culdesac roundabouts from 574136 candidates.
2025/01/25 19:43:04.284139 [INFO] Finished with 240311113 routable ways containing 2689869422 nodes
2025/01/25 19:43:05.144427 [INFO] Sorting osm access tags by way id...
2025/01/25 19:44:07.735169 [INFO] Finished
2025/01/25 19:44:10.217313 [INFO] Write OSMData to temp files
2025/01/25 19:44:14.069301 [INFO] Done
2025/01/25 19:44:14.083919 [INFO] Parsing files for relations: planet-250120.osm.pbf
2025/01/25 19:44:14.092733 [INFO] Parsing relations...
2025/01/25 21:04:43.699395 [INFO] invalid_argument thrown for condition  Jul 1- Aug 31 off
2025/01/25 21:05:37.157431 [INFO] Finished with 1854992 simple turn restrictions
2025/01/25 21:05:37.158242 [INFO] Finished with 0 lane connections
2025/01/25 21:05:37.176079 [INFO] Sorting complex restrictions by from id...
2025/01/25 21:05:37.265273 [INFO] Sorting complex restrictions by to id...
2025/01/25 21:05:37.354710 [INFO] Finished
2025/01/25 21:05:37.355217 [INFO] Write OSMData to temp files
2025/01/25 21:05:42.620149 [INFO] Done
2025/01/25 21:05:42.621220 [INFO] Parsing files for nodes: planet-250120.osm.pbf
2025/01/25 21:05:42.621244 [INFO] Sorting osm way node references by node id...
2025/01/25 21:48:41.305255 [INFO] Parsing nodes...
2025/01/25 23:51:30.008348 [INFO] Finished with 2334325010 nodes contained in routable ways
2025/01/25 23:51:30.009757 [INFO] Sorting osm way node references by way index and node shape index...
2025/01/26 00:36:36.142506 [INFO] Finished: changeset id 161544841
2025/01/26 00:36:36.144062 [INFO] Number of nodes with refs (exits) = 308779
2025/01/26 00:36:36.144069 [INFO] Number of nodes with exit_to = 3753
2025/01/26 00:36:36.144073 [INFO] Number of nodes with names = 349319
2025/01/26 00:36:36.144076 [INFO] Number of nodes with linguistics = 70395
2025/01/26 00:36:36.144080 [INFO] Number of way refs = 260724
2025/01/26 00:36:36.144083 [INFO] Number of reverse way refs = 18949
2025/01/26 00:36:36.144087 [INFO] Unique Node Strings (names, refs, etc.) = 146269
2025/01/26 00:36:36.144090 [INFO] Unique Strings (names, refs, etc.) = 11283306
2025/01/26 00:36:36.157147 [INFO] Write OSMData to temp files
2025/01/26 00:36:41.635163 [INFO] Done
```

</details>

<details>
<summary>Full build tiles log from PR</summary>

```
2025/01/25 15:07:20.281417 [INFO] Running valhalla_build_tiles with 8 thread(s).
…