@kevinkreiser 2025-02-04
just reporting back that we are saving about 1h45m (out of previously 6h) on our parsing staging with this update on the open servers